pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"
pillow = "^10.4.0"
markdown = "^3.7"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
coverage = "^7.6.1"
setuptools = "^74.0.0"
distribute = "^0.7.3"
django-cors-headers = "^4.4.0"
redis = "^5.0.8"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
eventlet = "^0.36.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
